move the InputField and SelectOptions component from the info_page_shared folder to the components folder as these components are not only info page specific
