+++ src/components/Select.js
@@ -4,7 +4,8 @@
 //     { value: 'fr', label: 'French' },
 //     { value: 'da', label: 'Danish' }
 //   ]
-import * as s from "../pages/info_page_shared/SelectOptions.sc";
+import * as s from "./SelectOptions.sc";
+
 export default function Select({
   elements,
   label,
