+++ src/pages/ResetPasswordStep2.js
@@ -7,7 +7,7 @@ import useFormField from "../hooks/useFormField";
 import Form from "./info_page_shared/Form";
 import FormSection from "./info_page_shared/FormSection";
 import FullWidthErrorMsg from "./info_page_shared/FullWidthErrorMsg";
-import InputField from "./info_page_shared/InputField";
+import InputField from "../components/InputField";
 import ButtonContainer from "./info_page_shared/ButtonContainer";
 import Button from "./info_page_shared/Button";
