+++ src/pages/settings_pages/ProfileDetails.js
@@ -8,7 +8,7 @@ import Form from "../info_page_shared/Form";
 import FormSection from "../info_page_shared/FormSection";
 import Button from "../info_page_shared/Button";
 import ButtonContainer from "../info_page_shared/ButtonContainer";
-import InputField from "../info_page_shared/InputField";
+import InputField from "../../components/InputField";
 import InfoPage from "../info_page_shared/InfoPage";
 import Header from "../info_page_shared/Header";
 import Heading from "../info_page_shared/Heading";
